Add utils/macros module
