+++ Cargo.toml
@@ -14,3 +14,4 @@ serde_json = "*"
 backtrace = "*"
 failure = "*"
 failure_derive = "*"
+heck = "*"
